Fix Manga MyRereading
This commit changes the previously wrong Manga struct tag my_rewatching
to my_rereadingg so that it is possible to get its value from the XML
response. The field is renamed accordingly and its type is also changed
from string to int. While bool might be more appropriate (since the only
possible values seem to be 1 and 0), int is used so that it matches with
the equivalent MangaEntry.EnableRereading field which is officially
documented to be an int type in the MyAnimeList API documentation.

It total this commit performs the following changes:

 * Anime.MyRewatching string -> Anime.MyRewatching int
 * Manga.MyRewatching string -> Manga.MyRereading  int
 * Manga.MyRewatchingEp      -> Manga.MyRereadingChap

Since by having the wrong tag it was previously impossible to get the
my_rereadingg value, this is a necessary breaking change.