Change AnimeEntry Status type from string to int
This commit changes AnimeEntry.Status and MangaEntry.Status from string
to int.

According to the official MyAnimeList API documentation the type of an
entry status can be either int or string. The initial thought was to use
the string type to allow the user to enter the value in any form they
want e.g. "watching" or "1". After using this package for a while, it
became obvious that it is more convenient and consistent to use the int
type (same as Anime.Status) which not only allows to use package
constants like StatusWatching as values but it also makes comparisons
between Anime.Status and AnimeEntry.Status easier. Same goes for the
equivalent Manga types.

The necessary breaking change in
90de691e933b6b98be5a71dcd4a5b0fbdfd8907d makes this change come easier.