Use functional options for NewClient
This commit changes the NewClient API to use functional options.
Previously nil had to always passed to NewClient except the rarer cases
when a custom http.Client was needed. The default case is now changed
from mal.NewClient(nil) to mal.NewClient() while providing a new option
HTTPClient to pass a custom http.Client when needed.

Additionally, the way the user provided credentials was through the
SetCredentials method. This method has been removed and replaced with
the Auth option which can only be passed when creating a new client.

The new API should be nicer to use while it remains open for extension.